Move convert Flag function in LanguageSelector
