+++ src/components/LanguageSelector.jsx
@@ -1,9 +1,33 @@
 import { Col, Row, Select } from "antd";
 import PropTypes from "prop-types";
-import convertFlag from "./convertFlag";
 
 
 const { Option } = Select
+
+function convertFlag(language) {
+    switch (language) {
+        case "German":
+            return "🇩🇪";
+        case "French":
+            return "🇲🇫";
+        case "Esperanto":
+            return "🏴‍☠️";
+        case "Spanish":
+            return "🇪🇦";
+        case "Russian":
+            return "🇷🇺";
+        case "Italian":
+            return "🇮🇹";
+        case "Japanese":
+            return "🇯🇵";
+        case "Portuguese":
+            return "🇵🇹";
+        case "Dutch":
+            return "🇳🇱";
+        default:
+            return "🇬🇧";
+    }
+}
 
 const LanguageSelector = ({ languages, language, onChange }) => {
     // Build list of language items, alphabetically sorted
